# riskcast-v16-main/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RISKCAST V3 API
  description: |
    Enterprise Insurance-Grade Risk Intelligence Platform API
    
    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.
    
    ## Versioning
    API is versioned via URL path (e.g., `/api/v3/...`)
    
    ## Rate Limiting
    - Standard: 100 requests/minute
    - Authenticated: 1000 requests/minute
    
    ## Audit Trail
    All write operations are recorded in an immutable audit ledger with hash chain verification.
    
    ## Multi-Tenancy
    All operations are automatically scoped to the authenticated user's tenant.
    
    ## Permissions
    Endpoints require specific RBAC permissions. See individual endpoint documentation.
  version: 3.0.0
  contact:
    name: API Support
    email: api-support@riskcast.com
  license:
    name: Proprietary
    url: https://riskcast.com/terms
servers:
  - url: https://api.riskcast.com/api/v3
    description: Production
  - url: https://staging-api.riskcast.com/api/v3
    description: Staging
  - url: http://localhost:8000/api/v3
    description: Local Development
...
paths:
...
  /policies/{policy_id}/decision-pack:
    get:
      tags: [Policies, Compliance]
      summary: Download decision pack
      description: |
        Download the complete decision pack for regulatory compliance.
        
        Includes:
        - Policy details
        - Quote and pricing
        - Risk assessment
        - Model version
        - Evidence bundle manifest
        - Full audit trail
        
        **Permission Required**: `COMPLIANCE_READ`
      operationId: downloadDecisionPack
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PolicyId'
      responses:
        '200':
          description: Decision pack ZIP file
          content:
            application/zip:
              schema:
                type: string
                format: binary

  # ===========================================================================
  # Claims
  # ===========================================================================
...
  /compliance/gdpr/export:
    post:
      tags: [Compliance]
      summary: Export user data (DSAR)
      description: |
        Export all user data for Data Subject Access Request.
        
        **Permission Required**: `COMPLIANCE_READ`
      operationId: exportUserData
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  format: uuid
              required: [user_id]
      responses:
        '200':
          description: User data export ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /compliance/gdpr/delete:
    post:
      tags: [Compliance]
      summary: Delete user data
      description: |
        Process deletion request (Right to Erasure).
        
        Note: Some data may be retained for legal/regulatory reasons.
        
        **Permission Required**: `COMPLIANCE_WRITE`
      operationId: deleteUserData
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  format: uuid
              required: [user_id]
      responses:
        '200':
          description: Deletion report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionReport'
components:
...
  schemas:
...
    DeletionReport:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        deleted_at:
          type: string
          format: date-time
        records_deleted:
          type: integer
        records_retained:
          type: integer
        retention_reason:
          type: string
...
  parameters:
...
    PolicyId:
      name: policy_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Policy ID
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from authentication endpoint.
        Include in Authorization header as: `Bearer <token>`
security:
  - BearerAuth: []